pyproject.toml
[project]
name = "rag-demo"
version = "0.1.0"
description = "A RAG application with document processing and Q&A capabilities"
authors = [
    {name = "RAG Demo", email = "demo@example.com"}
]
requires-python = ">=3.10"
dependencies = [
    "dspy-ai>=2.4.0",
    "qdrant-client>=1.7.0",
    "openai>=1.0.0",
    "streamlit>=1.28.0",
    "pypdf2>=3.0.1",
    "python-docx>=0.8.11",
    "python-pptx>=0.6.21",
    "python-multipart>=0.0.6",
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "numpy>=1.24.0",
    "loguru>=0.7.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pytest-cov>=6.2.1",
]
